Pull LocalH2 and Persistence Unit out of DatamartExporter
